Safari-611.1.10/LayoutTests/fast/viewport/viewport-44-expected.txt

ALERT: viewport size 10000x10000 scale 0.25 with limits [0.25, 5] and userScalable true
